database and changelog organization
