+++ modules/consensus/database.go
@@ -1,23 +1,14 @@
 package consensus
 
-// database.go contains initialization functions for the database, and helper
-// functions for accessing the database. All database access functions take a
-// bolt.Tx as input because consensus manipulations should always be made as a
-// single atomic transaction. Any function for changing the database will not
-// return errors, but instead will panic as a sanity check. No item should ever
-// be inserted into the database that is already in the database, and no item
-// should ever be removed from the database that is not currently in the
-// database. Attempting such with the debug flags enabled indicate developer
-// error and will cause a panic.
+// database.go contains functions to initialize the database and report
+// inconsistencies. All of the database-specific logic belongs here.
 
 import (
 	"errors"
 
 	"github.com/NebulousLabs/Sia/build"
 	"github.com/NebulousLabs/Sia/encoding"
-	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/persist"
-	"github.com/NebulousLabs/Sia/types"
 
 	"github.com/NebulousLabs/bolt"
 )
@@ -29,55 +20,10 @@ var (
 	errDBInconsistent = errors.New("database guard indicates inconsistency within database")
 	errNonEmptyBucket = errors.New("cannot remove a map with objects still in it")
 
-	prefixDSCO = []byte("dsco_")
-	prefixFCEX = []byte("fcex_")
-
 	dbMetadata = persist.Metadata{
 		Version: "0.4.3",
 		Header:  "Consensus Set Database",
 	}
-
-	// BlockHeight is a bucket that stores the current block height.
-	//
-	// Generally we would just look at BlockPath.Stats(), but there is an error
-	// in boltdb that prevents the bucket stats from updating until a tx is
-	// committed. Wasn't a problem until we started doing the entire block as
-	// one tx.
-	//
-	// DEPRECATED - block.Stats() should be sufficient to determine the block
-	// height, but currently stats are only computed after committing a
-	// transaction, therefore cannot be assumed reliable.
-	BlockHeight = []byte("BlockHeight")
-
-	// BlockMap is a database bucket containing all of the processed blocks,
-	// keyed by their id. This includes blocks that are not currently in the
-	// consensus set, and blocks that may not have been fully validated yet.
-	BlockMap = []byte("BlockMap")
-
-	// BlockPath is a database bucket containing a mapping from the height of a
-	// block to the id of the block at that height. BlockPath only includes
-	// blocks in the current path.
-	BlockPath = []byte("BlockPath")
-
-	// Consistency is a database bucket with a flag indicating whether
-	// inconsistencies within the database have been detected.
-	Consistency = []byte("Consistency")
-
-	// SiacoinOutputs is a database bucket that contains all of the unspent
-	// siacoin outputs.
-	SiacoinOutputs = []byte("SiacoinOutputs")
-
-	// FileContracts is a database bucket that contains all of the open file
-	// contracts.
-	FileContracts = []byte("FileContracts")
-
-	// SiafundOutputs is a database bucket that contains all of the unspent
-	// siafund outputs.
-	SiafundOutputs = []byte("SiafundOutputs")
-
-	// SiafundPool is a database bucket storing the current value of the
-	// siafund pool.
-	SiafundPool = []byte("SiafundPool")
 )
 
 type (
@@ -112,80 +58,32 @@ func (cs *ConsensusSet) openDB(filename string) (err error) {
 }
 
 // dbInitialized returns true if the database appears to be initialized, false
-// if not.
+// if not. Checking for the existence of the siafund pool bucket is typically
+// sufficient to determine whether the database has gone through the
+// initialization process.
 func dbInitialized(tx *bolt.Tx) bool {
-	// If the SiafundPool bucket exists, the database has almost certainly been
-	// initialized correctly.
 	return tx.Bucket(SiafundPool) != nil
 }
 
 // initDatabase is run when the database. This has become the true
 // init function for consensus set
 func (cs *ConsensusSet) initDB(tx *bolt.Tx) error {
-	// Create the change log - needs to hapen before any sort of subscriber
-	// updates are made.
-	createChangeLog(tx)
-
-	// Enumerate the database buckets.
-	buckets := [][]byte{
-		BlockHeight,
-		BlockMap,
-		BlockPath,
-		Consistency,
-		SiacoinOutputs,
-		FileContracts,
-		SiafundOutputs,
-		SiafundPool,
+	// Create the compononents of the database.
+	err := cs.createConsensusDB(tx)
+	if err != nil {
+		return err
 	}
-
-	// Create the database buckets.
-	for _, bucket := range buckets {
-		_, err := tx.CreateBucket(bucket)
-		if err != nil {
-			return err
-		}
+	err = createChangeLog(tx)
+	if err != nil {
+		return err
 	}
 
 	// Place a 'false' in the consistency bucket to indicate that no
 	// inconsistencies have been found.
-	err := tx.Bucket(Consistency).Put(Consistency, encoding.Marshal(false))
+	err = tx.Bucket(Consistency).Put(Consistency, encoding.Marshal(false))
 	if err != nil {
 		return err
 	}
-
-	// Set the block height to -1, so the genesis block is at height 0.
-	blockHeight := tx.Bucket(BlockHeight)
-	underflow := types.BlockHeight(0)
-	err = blockHeight.Put(BlockHeight, encoding.Marshal(underflow-1))
-	if err != nil {
-		return err
-	}
-
-	// Set the siafund pool to 0.
-	setSiafundPool(tx, types.NewCurrency64(0))
-
-	// Update the siafund output diffs map for the genesis block on disk. This
-	// needs to happen between the database being opened/initilized and the
-	// consensus set hash being calculated
-	for _, sfod := range cs.blockRoot.SiafundOutputDiffs {
-		commitSiafundOutputDiff(tx, sfod, modules.DiffApply)
-	}
-
-	// Add the miner payout from the genesis block to the delayed siacoin
-	// outputs - unspendable, as the unlock hash is blank.
-	createDSCOBucket(tx, types.MaturityDelay)
-	addDSCO(tx, types.MaturityDelay, cs.blockRoot.Block.MinerPayoutID(0), types.SiacoinOutput{
-		Value:      types.CalculateCoinbase(0),
-		UnlockHash: types.UnlockHash{},
-	})
-
-	// Add the genesis block to the block strucutres - checksum must be taken
-	// after pushing the genesis block into the path.
-	pushPath(tx, cs.blockRoot.Block.ID())
-	if build.DEBUG {
-		cs.blockRoot.ConsensusChecksum = consensusChecksum(tx)
-	}
-	addBlockMap(tx, &cs.blockRoot)
 	return nil
 }
 
@@ -210,402 +108,4 @@ func markInconsistency(tx *bolt.Tx) {
 		panic(err)
 	}
 
-}
-
-// blockHeight returns the height of the blockchain.
-func blockHeight(tx *bolt.Tx) types.BlockHeight {
-	var height types.BlockHeight
-	bh := tx.Bucket(BlockHeight)
-	err := encoding.Unmarshal(bh.Get(BlockHeight), &height)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-
-	// Check that there was not an underflow on the height.
-	zeroHeight := types.BlockHeight(0)
-	if height > zeroHeight-1e9 {
-		panic(height)
-	}
-	return types.BlockHeight(height)
-}
-
-// currentBlockID returns the id of the most recent block in the consensus set.
-func currentBlockID(tx *bolt.Tx) types.BlockID {
-	id, err := getPath(tx, blockHeight(tx))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	return id
-}
-
-// currentProcessedBlock returns the most recent block in the consensus set.
-func currentProcessedBlock(tx *bolt.Tx) *processedBlock {
-	pb, err := getBlockMap(tx, currentBlockID(tx))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	return pb
-}
-
-// getBlockMap returns a processed block with the input id.
-func getBlockMap(tx *bolt.Tx, id types.BlockID) (*processedBlock, error) {
-	// Look up the encoded block.
-	pbBytes := tx.Bucket(BlockMap).Get(id[:])
-	if pbBytes == nil {
-		return nil, errNilItem
-	}
-
-	// Decode the block - should never fail.
-	var pb processedBlock
-	err := encoding.Unmarshal(pbBytes, &pb)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	return &pb, nil
-}
-
-// addBlockMap adds a processed block to the block map.
-func addBlockMap(tx *bolt.Tx, pb *processedBlock) {
-	id := pb.Block.ID()
-	err := tx.Bucket(BlockMap).Put(id[:], encoding.Marshal(*pb))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// getPath returns the block id at 'height' in the block path.
-func getPath(tx *bolt.Tx, height types.BlockHeight) (id types.BlockID, err error) {
-	idBytes := tx.Bucket(BlockPath).Get(encoding.Marshal(height))
-	if idBytes == nil {
-		return types.BlockID{}, errNilItem
-	}
-
-	err = encoding.Unmarshal(idBytes, &id)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	return id, nil
-}
-
-// pushPath adds a block to the BlockPath at current height + 1.
-func pushPath(tx *bolt.Tx, bid types.BlockID) {
-	// Fetch and update the block height.
-	bh := tx.Bucket(BlockHeight)
-	heightBytes := bh.Get(BlockHeight)
-	var oldHeight types.BlockHeight
-	err := encoding.Unmarshal(heightBytes, &oldHeight)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	newHeightBytes := encoding.Marshal(oldHeight + 1)
-	err = bh.Put(BlockHeight, newHeightBytes)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-
-	// Add the block to the block path.
-	bp := tx.Bucket(BlockPath)
-	err = bp.Put(newHeightBytes, bid[:])
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// popPath removes a block from the "end" of the chain, i.e. the block
-// with the largest height.
-func popPath(tx *bolt.Tx) {
-	// Fetch and update the block height.
-	bh := tx.Bucket(BlockHeight)
-	oldHeightBytes := bh.Get(BlockHeight)
-	var oldHeight types.BlockHeight
-	err := encoding.Unmarshal(oldHeightBytes, &oldHeight)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	newHeightBytes := encoding.Marshal(oldHeight - 1)
-	err = bh.Put(BlockHeight, newHeightBytes)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-
-	// Remove the block from the path - make sure to remove the block at
-	// oldHeight.
-	bp := tx.Bucket(BlockPath)
-	err = bp.Delete(oldHeightBytes)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// isSiacoinOutput returns true if there is a siacoin output of that id in the
-// database.
-func isSiacoinOutput(tx *bolt.Tx, id types.SiacoinOutputID) bool {
-	bucket := tx.Bucket(SiacoinOutputs)
-	sco := bucket.Get(id[:])
-	return sco != nil
-}
-
-// getSiacoinOutput fetches a siacoin output from the database. An error is
-// returned if the siacoin output does not exist.
-func getSiacoinOutput(tx *bolt.Tx, id types.SiacoinOutputID) (types.SiacoinOutput, error) {
-	scoBytes := tx.Bucket(SiacoinOutputs).Get(id[:])
-	if scoBytes == nil {
-		return types.SiacoinOutput{}, errNilItem
-	}
-	var sco types.SiacoinOutput
-	err := encoding.Unmarshal(scoBytes, &sco)
-	if err != nil {
-		return types.SiacoinOutput{}, err
-	}
-	return sco, nil
-}
-
-// addSiacoinOutput adds a siacoin output to the database. An error is returned
-// if the siacoin output is already in the database.
-func addSiacoinOutput(tx *bolt.Tx, id types.SiacoinOutputID, sco types.SiacoinOutput) {
-	// While this is not supposed to be allowed, there's a bug in the consensus
-	// code which means that earlier versions have accetped 0-value outputs
-	// onto the blockchain. A hardfork to remove 0-value outputs will fix this,
-	// and that hardfork is planned, but not yet.
-	/*
-		if build.DEBUG && sco.Value.IsZero() {
-			panic("discovered a zero value siacoin output")
-		}
-	*/
-	siacoinOutputs := tx.Bucket(SiacoinOutputs)
-	// Sanity check - should not be adding an item that exists.
-	if build.DEBUG && siacoinOutputs.Get(id[:]) != nil {
-		panic("repeat siacoin output")
-	}
-	err := siacoinOutputs.Put(id[:], encoding.Marshal(sco))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// removeSiacoinOutput removes a siacoin output from the database. An error is
-// returned if the siacoin output is not in the database prior to removal.
-func removeSiacoinOutput(tx *bolt.Tx, id types.SiacoinOutputID) {
-	scoBucket := tx.Bucket(SiacoinOutputs)
-	// Sanity check - should not be removing an item that is not in the db.
-	if build.DEBUG && scoBucket.Get(id[:]) == nil {
-		panic("nil siacoin output")
-	}
-	err := scoBucket.Delete(id[:])
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// getFileContract fetches a file contract from the database, returning an
-// error if it is not there.
-func getFileContract(tx *bolt.Tx, id types.FileContractID) (fc types.FileContract, err error) {
-	fcBytes := tx.Bucket(FileContracts).Get(id[:])
-	if fcBytes == nil {
-		return types.FileContract{}, errNilItem
-	}
-	err = encoding.Unmarshal(fcBytes, &fc)
-	if err != nil {
-		return types.FileContract{}, err
-	}
-	return fc, nil
-}
-
-// addFileContract adds a file contract to the database. An error is returned
-// if the file contract is already in the database.
-func addFileContract(tx *bolt.Tx, id types.FileContractID, fc types.FileContract) {
-	// Add the file contract to the database.
-	fcBucket := tx.Bucket(FileContracts)
-	// Sanity check - should not be adding a zero-payout file contract.
-	if build.DEBUG && fc.Payout.IsZero() {
-		panic("adding zero-payout file contract")
-	}
-	// Sanity check - should not be adding a file contract already in the db.
-	if build.DEBUG && fcBucket.Get(id[:]) != nil {
-		panic("repeat file contract")
-	}
-	err := fcBucket.Put(id[:], encoding.Marshal(fc))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-
-	// Add an entry for when the file contract expires.
-	expirationBucketID := append(prefixFCEX, encoding.Marshal(fc.WindowEnd)...)
-	expirationBucket, err := tx.CreateBucketIfNotExists(expirationBucketID)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	err = expirationBucket.Put(id[:], []byte{})
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// removeFileContract removes a file contract from the database.
-func removeFileContract(tx *bolt.Tx, id types.FileContractID) {
-	// Delete the file contract entry.
-	fcBucket := tx.Bucket(FileContracts)
-	fcBytes := fcBucket.Get(id[:])
-	// Sanity check - should not be removing a file contract not in the db.
-	if build.DEBUG && fcBytes == nil {
-		panic("nil file contract")
-	}
-	err := fcBucket.Delete(id[:])
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-
-	// Delete the entry for the file contract's expiration. The portion of
-	// 'fcBytes' used to determine the expiration bucket id is the
-	// byte-representation of the file contract window end, which always
-	// appears at bytes 48-56.
-	expirationBucketID := append(prefixFCEX, fcBytes[48:56]...)
-	expirationBucket := tx.Bucket(expirationBucketID)
-	expirationBytes := expirationBucket.Get(id[:])
-	if expirationBytes == nil {
-		panic(errNilItem)
-	}
-	err = expirationBucket.Delete(id[:])
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// getSiafundOutput fetches a siafund output from the database. An error is
-// returned if the siafund output does not exist.
-func getSiafundOutput(tx *bolt.Tx, id types.SiafundOutputID) (types.SiafundOutput, error) {
-	sfoBytes := tx.Bucket(SiafundOutputs).Get(id[:])
-	if sfoBytes == nil {
-		return types.SiafundOutput{}, errNilItem
-	}
-	var sfo types.SiafundOutput
-	err := encoding.Unmarshal(sfoBytes, &sfo)
-	if err != nil {
-		return types.SiafundOutput{}, err
-	}
-	return sfo, nil
-}
-
-// addSiafundOutput adds a siafund output to the database. An error is returned
-// if the siafund output is already in the database.
-func addSiafundOutput(tx *bolt.Tx, id types.SiafundOutputID, sfo types.SiafundOutput) {
-	siafundOutputs := tx.Bucket(SiafundOutputs)
-	// Sanity check - should not be adding a siafund output with a value of
-	// zero.
-	if build.DEBUG && sfo.Value.IsZero() {
-		panic("zero value siafund being added")
-	}
-	// Sanity check - should not be adding an item already in the db.
-	if build.DEBUG && siafundOutputs.Get(id[:]) != nil {
-		panic("repeat siafund output")
-	}
-	err := siafundOutputs.Put(id[:], encoding.Marshal(sfo))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// removeSiafundOutput removes a siafund output from the database. An error is
-// returned if the siafund output is not in the database prior to removal.
-func removeSiafundOutput(tx *bolt.Tx, id types.SiafundOutputID) {
-	sfoBucket := tx.Bucket(SiafundOutputs)
-	if build.DEBUG && sfoBucket.Get(id[:]) == nil {
-		panic("nil siafund output")
-	}
-	err := sfoBucket.Delete(id[:])
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// getSiafundPool returns the current value of the siafund pool. No error is
-// returned as the siafund pool should always be available.
-func getSiafundPool(tx *bolt.Tx) (pool types.Currency) {
-	bucket := tx.Bucket(SiafundPool)
-	poolBytes := bucket.Get(SiafundPool)
-	// An error should only be returned if the object stored in the siafund
-	// pool bucket is either unavailable or otherwise malformed. As this is a
-	// developer error, a panic is appropriate.
-	err := encoding.Unmarshal(poolBytes, &pool)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-	return pool
-}
-
-// setSiafundPool updates the saved siafund pool on disk
-func setSiafundPool(tx *bolt.Tx, c types.Currency) {
-	err := tx.Bucket(SiafundPool).Put(SiafundPool, encoding.Marshal(c))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// addDSCO adds a delayed siacoin output to the consnesus set.
-func addDSCO(tx *bolt.Tx, bh types.BlockHeight, id types.SiacoinOutputID, sco types.SiacoinOutput) {
-	// Sanity check - dsco should never have a value of zero.
-	// An error in the consensus code means sometimes there are 0-value dscos
-	// in the blockchain. A hardfork will fix this.
-	/*
-		if build.DEBUG && sco.Value.IsZero() {
-			panic("zero-value dsco being added")
-		}
-	*/
-	// Sanity check - output should not already be in the full set of outputs.
-	if build.DEBUG && tx.Bucket(SiacoinOutputs).Get(id[:]) != nil {
-		panic("dsco already in output set")
-	}
-	dscoBucketID := append(prefixDSCO, encoding.EncUint64(uint64(bh))...)
-	dscoBucket := tx.Bucket(dscoBucketID)
-	// Sanity check - should not be adding an item already in the db.
-	if build.DEBUG && dscoBucket.Get(id[:]) != nil {
-		panic(errRepeatInsert)
-	}
-	err := dscoBucket.Put(id[:], encoding.Marshal(sco))
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// removeDSCO removes a delayed siacoin output from the consensus set.
-func removeDSCO(tx *bolt.Tx, bh types.BlockHeight, id types.SiacoinOutputID) {
-	bucketID := append(prefixDSCO, encoding.Marshal(bh)...)
-	// Sanity check - should not remove an item not in the db.
-	dscoBucket := tx.Bucket(bucketID)
-	if build.DEBUG && dscoBucket.Get(id[:]) == nil {
-		panic("nil dsco")
-	}
-	err := dscoBucket.Delete(id[:])
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// createDSCOBucket creates a bucket for the delayed siacoin outputs at the
-// input height.
-func createDSCOBucket(tx *bolt.Tx, bh types.BlockHeight) {
-	bucketID := append(prefixDSCO, encoding.Marshal(bh)...)
-	_, err := tx.CreateBucket(bucketID)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
-}
-
-// deleteDSCOBucket deletes the bucket that held a set of delayed siacoin
-// outputs.
-func deleteDSCOBucket(tx *bolt.Tx, bh types.BlockHeight) {
-	// Delete the bucket.
-	bucketID := append(prefixDSCO, encoding.Marshal(bh)...)
-	bucket := tx.Bucket(bucketID)
-	if build.DEBUG && bucket == nil {
-		panic(errNilBucket)
-	}
-
-	// TODO: Check that the bucket is empty. Using Stats() does not work at the
-	// moment, as there is an error in the boltdb code.
-
-	err := tx.DeleteBucket(bucketID)
-	if build.DEBUG && err != nil {
-		panic(err)
-	}
 }
